New version 4.5.1 sir: added option -add_iterations, plus automatic recovery procedure

+++ test/system/sir.t
@@ -19,6 +19,8 @@ my $model_dir = $includes::testfiledir;
 my @commands = 
 	(
 	 get_command('sir') . " $model_dir/pheno.mod -samples=50,100 -resamples=25,50 -no-boxcox -dir=$dir", 
+	 get_command('sir') . "  -dir=$dir", 
+	 get_command('sir') . " -samples=50 -resamples=25 -add_iterations -dir=$dir", 
 	 get_command('sir') . " $model_dir/pheno.mod -samples=50 -resamples=25 -auto_rawres=0.9 -seed=50032 -dir=$dir",
 	 get_command('sir') . " $model_dir/sir/localmin.mod -samples=50,100 -resamples=25,50 -dir=$dir",
 	 get_command('sir') . " $model_dir/pheno.mod -samples=50,100,100 -resamples=25,50,50 -covmat_input=$model_dir/pheno_fake.cov -dir=$dir",
@@ -30,12 +32,14 @@ my @commands =
 	 get_command('sir') . " $model_dir/mox_sir.mod -samples=50 -resamples=100 -with_replacement -dir=$dir"
 	);
 
-foreach my $command (@commands){
+
+for (my $i=0; $i<scalar(@commands); $i++){
+	my $command = $commands[$i];
 	print "Running $command\n";
 	my $rc = system($command);
 	$rc = $rc >> 8;
 	ok ($rc == 0, "$command, should run ok");
-	rmtree([$dir]);
+	rmtree([$dir]) if ($i>1);
 }
 
 remove_test_dir($tempdir);
